fix interface; move expression functions and extractors to common package
