+++ core/dnsserver/view.go
@@ -1,8 +1,6 @@
 package dnsserver
 
 import (
-	"context"
-
 	"github.com/coredns/coredns/request"
 )
 
@@ -11,5 +9,5 @@ import (
 // serves a DNS request, it will route the request to the first Config (server block) that passes
 // all its filter functions.
 type Viewer interface {
-	Filter(ctx context.Context, req request.Request) bool
+	Filter(req request.Request) bool
 }
