+++ plugin/view/view.go
@@ -3,11 +3,12 @@ package view
 import (
 	"github.com/Knetic/govaluate"
 	"github.com/coredns/coredns/plugin"
+	"github.com/coredns/coredns/plugin/pkg/expression"
 )
 
 type View struct {
 	rules      []*govaluate.EvaluableExpression
-	extractors extractorMap
+	extractors expression.ExtractorMap
 	Next       plugin.Handler
 }
